+++ test/integration/scheduler_perf/BUILD
@@ -15,6 +15,7 @@ go_library(
     importpath = "k8s.io/kubernetes/test/integration/scheduler_perf",
     deps = [
         "//staging/src/k8s.io/api/core/v1:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/labels:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//staging/src/k8s.io/client-go/informers/core/v1:go_default_library",
@@ -23,6 +24,7 @@ go_library(
         "//staging/src/k8s.io/component-base/metrics/legacyregistry:go_default_library",
         "//staging/src/k8s.io/component-base/metrics/testutil:go_default_library",
         "//test/integration/util:go_default_library",
+        "//test/utils:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
 )
@@ -32,7 +34,7 @@ go_test(
     size = "large",
     srcs = [
         "main_test.go",
-        "scheduler_bench_test.go",
+        "scheduler_perf_legacy_test.go",
         "scheduler_perf_test.go",
         "scheduler_test.go",
     ],
